refactor: move subscription.hpp from odometry/ to utils/
