+++ px4_ros2_cpp/include/px4_ros2/odometry/angular_velocity.hpp
@@ -8,7 +8,7 @@
 #include <Eigen/Eigen>
 #include <px4_msgs/msg/vehicle_angular_velocity.hpp>
 #include <px4_ros2/common/context.hpp>
-#include <px4_ros2/odometry/subscription.hpp>
+#include <px4_ros2/utils/subscription.hpp>
 
 namespace px4_ros2
 {
