+++ px4_ros2_cpp/include/px4_ros2/odometry/attitude.hpp
@@ -8,8 +8,8 @@
 #include <Eigen/Eigen>
 #include <px4_msgs/msg/vehicle_attitude.hpp>
 #include <px4_ros2/common/context.hpp>
-#include <px4_ros2/odometry/subscription.hpp>
 #include <px4_ros2/utils/geometry.hpp>
+#include <px4_ros2/utils/subscription.hpp>
 
 namespace px4_ros2
 {
