+++ px4_ros2_cpp/include/px4_ros2/vehicle_state/vtol_status.hpp
@@ -7,7 +7,7 @@
 
 #include <px4_msgs/msg/vtol_vehicle_status.hpp>
 #include <px4_ros2/common/context.hpp>
-#include <px4_ros2/odometry/subscription.hpp>
+#include <px4_ros2/utils/subscription.hpp>
 
 namespace px4_ros2
 {
